+++ include/inttypes.h
@@ -15,11 +15,11 @@
 #define _INTTYPES_H_ 1
 
 typedef long _intptr_t;
-typedef unsigned long _uintptr_t;
+typedef unsigned long  _uintptr_t;
 
 typedef _intptr_t intptr_t;
 typedef _uintptr_t uintptr_t;
-typedef long long	intmax_t;
-typedef unsigned long long	uintmax_t;
+typedef long	intmax_t;
+typedef unsigned long	uintmax_t;
 
 #endif
