+++ lib/ap/plan9/malloc.c
@@ -36,9 +36,9 @@ struct Arena
 };
 static Arena arena;
 
-#define datoff		((int64_t)((Bucket*)0)->data)
+#define datoff		((size_t)((Bucket*)0)->data)
 
-extern	void	*sbrk(unsigned long);
+extern	void	*sbrk(uint32_t);
 
 void*
 malloc(size_t size)
