+++ lib/ap/plan9/times.c
@@ -15,6 +15,7 @@
 #include <errno.h>
 #include <string.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 static
 char*
@@ -33,7 +34,7 @@ times(struct tms *buf)
 {
 	char b[200], *p;
 	int f;
-	unsigned long r;
+	uint32_t r;
 
 	memset(b, 0, sizeof(b));
 	f = open("/dev/cputime", O_RDONLY);
